[DROOLS-5643] Separate DMN PMML tests to a separate module (#2258)
