+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kie-server-testing-client-deployment-settings.xml
@@ -1,0 +1,25 @@
+<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd">
+  <!-- Using the remote repo dir here, so the kie-ci deploys the artifacts into this one, instead of the default local
+       repo. FIXME this will break once the client (test class) is not on same machine as the remote repo -->
+  <localRepository>${kie.server.testing.remote.repo.dir}</localRepository>
+
+  ${maven.settings.mirror}
+
+  <profiles>
+    <profile>
+      <id>additional-maven-repos</id>
+      <repositories>
+        <repository>
+          <id>remote-testing-repo</id>
+          <url>${kie.server.testing.remote.repo.url}</url>
+        </repository>
+      </repositories>
+    </profile>
+  </profiles>
+
+  <activeProfiles>
+    <activeProfile>additional-maven-repos</activeProfile>
+  </activeProfiles>
+</settings>
