+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kie-server-testing-server-custom-settings.xml
@@ -1,0 +1,26 @@
+<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0 http://maven.apache.org/xsd/settings-1.0.0.xsd">
+  <!-- Dividing the local Maven repository on the server and the remote repository with kjars closely simulates the
+       real-world deployments where the remote repo is typically located on different machine. -->
+  <localRepository>${kie.server.testing.server.local.repo.dir}</localRepository>
+
+  ${maven.settings.mirror}
+
+  <profiles>
+    <profile>
+      <id>additional-maven-repos</id>
+      <repositories>
+        <repository>
+          <id>remote-testing-repo</id>
+          <url>${kie.server.testing.remote.repo.url}</url>
+        </repository>
+      </repositories>
+    </profile>
+  </profiles>
+
+  <activeProfiles>
+    <activeProfile>additional-maven-repos</activeProfile>
+  </activeProfiles>
+
+</settings>
