+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kjars-sources/dmn-pmml-kmeans/src/main/resources/KMeans.dmn
