+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kjars-sources/dmn-pmml-mining/src/main/resources/SumMiningModel.dmn
