+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kjars-sources/dmn-pmml-naive-bayes/src/main/resources/META-INF/kmodule.xml
