+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kjars-sources/dmn-pmml-nn/src/main/resources/test_nn.pmml
