+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/filtered-resources/kjars-sources/dmn-pmml-scorecard/src/main/resources/KiePMMLScoreCard.dmn
