+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/java/org/kie/server/integrationtests/dmn/pmml/DMNPMMLRegressionIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,12 @@
  * limitations under the License.
  */
 
+
 package org.kie.server.integrationtests.dmn.pmml;
 
-import java.util.Map;
+import java.math.BigDecimal;
 
 import org.assertj.core.api.Assertions;
-import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.kie.api.KieServices;
@@ -29,38 +29,39 @@ import org.kie.server.api.model.KieServiceResponse;
 import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.ServiceResponse;
 import org.kie.server.client.KieServicesClient;
-import org.kie.server.integrationtests.dmn.DMNKieServerBaseIntegrationTest;
 import org.kie.server.integrationtests.shared.KieServerAssert;
 import org.kie.server.integrationtests.shared.KieServerDeployer;
 
-public class DMNPMMLKMeansIntegrationTest extends DMNKieServerBaseIntegrationTest {
+/*
+ * This it.test is reportedly not working in the "embedded" EE container,
+ * working correctly instead with the proper EE container activated with mvn profiles ("wildfly", etc.)
+ */
+public class DMNPMMLRegressionIntegrationTest extends DMNPMMLKieServerBaseIntegrationTest {
 
     private static final ReleaseId kjar1 = new ReleaseId(
-            "org.kie.server.testing", "dmn-pmml-kmeans-kjar",
+            "org.kie.server.testing", "dmn-pmml-regression-kjar",
             "1.0.0.Final" );
 
-    private static final String CONTAINER_1_ID  = "dmn-pml-kmeans";
+    private static final String CONTAINER_1_ID  = "dmn-pmml-regression";
 
-    private static final String KMEANS_MODEL_NAMESPACE
+    private static final String REGRESSION_MODEL_NAMESPACE
             = "https://kiegroup.org/dmn/_51A1FD67-8A67-4332-9889-B718BE8B7456";
-    private static final String KMEANS_MODEL_NAME = "KMeansDMN";
-    private static final String KMEANS_DECISION_NAME = "Decision1";
+    private static final String REGRESSION_MODEL_NAME = "TestRegressionDMN";
+    private static final String REGRESSION_DECISION_NAME = "Decision";
 
     private static final long EXTENDED_TIMEOUT = 300000L;
 
     @BeforeClass
     public static void deployArtifacts() {
-        if (DMNPMMLTestUtils.extendedPMMLTestsEnabled() == true) {
-            KieServerDeployer.buildAndDeployCommonMavenParent();
-            KieServerDeployer.buildAndDeployMavenProjectFromResource("/kjars-sources/dmn-pmml-kmeans");
+        KieServerDeployer.buildAndDeployCommonMavenParent();
+        KieServerDeployer.buildAndDeployMavenProjectFromResource("/kjars-sources/dmn-pmml-regression");
 
-            kieContainer = KieServices.Factory.get().newKieContainer(kjar1);
+        kieContainer = KieServices.Factory.get().newKieContainer(kjar1);
 
-            // Having timeout issues due to pmml -> raised timeout.
-            KieServicesClient client = createDefaultStaticClient(EXTENDED_TIMEOUT);
-            ServiceResponse<KieContainerResource> reply = client.createContainer(CONTAINER_1_ID, new KieContainerResource(CONTAINER_1_ID, kjar1));
-            KieServerAssert.assertSuccess(reply);
-        }
+        // Having timeout issues due to pmml -> raised timeout.
+        KieServicesClient client = createDefaultStaticClient(EXTENDED_TIMEOUT);
+        ServiceResponse<KieContainerResource> reply = client.createContainer(CONTAINER_1_ID, new KieContainerResource(CONTAINER_1_ID, kjar1));
+        KieServerAssert.assertSuccess(reply);
     }
 
     /*
@@ -68,23 +69,24 @@ public class DMNPMMLKMeansIntegrationTest extends DMNKieServerBaseIntegrationTes
      * working correctly instead with the proper EE container activated with mvn profiles ("wildfly", etc.)
      */
     @Test
-    public void testDMNWithPMMLKmeans() {
-        Assume.assumeTrue(DMNPMMLTestUtils.extendedPMMLTestsEnabled());
-
+    public void testDMNWithPMMLRegression() {
         final DMNContext dmnContext = dmnClient.newContext();
-        dmnContext.set("x", 5);
-        dmnContext.set("y", 5);
-
+        dmnContext.set("fld1", 3);
+        dmnContext.set("fld2", 2);
+        dmnContext.set("fld3", "y");
         final ServiceResponse<DMNResult> serviceResponse = dmnClient.evaluateDecisionByName(
                 CONTAINER_1_ID,
-                KMEANS_MODEL_NAMESPACE,
-                KMEANS_MODEL_NAME,
-                KMEANS_DECISION_NAME,
+                REGRESSION_MODEL_NAMESPACE,
+                REGRESSION_MODEL_NAME,
+                REGRESSION_DECISION_NAME,
                 dmnContext);
 
         Assertions.assertThat(serviceResponse.getType()).isEqualTo(KieServiceResponse.ResponseType.SUCCESS);
-        Assertions.assertThat(serviceResponse.getResult().getDecisionResultByName("Decision1").getResult()).isNotNull();
-        final Map<String, Object> decisionResult = (Map<String, Object>) serviceResponse.getResult().getDecisionResultByName("Decision1").getResult();
-        Assertions.assertThat((String) decisionResult.get("predictedValue")).isEqualTo("4");
+
+        final DMNResult dmnResult = serviceResponse.getResult();
+        Assertions.assertThat(dmnResult).isNotNull();
+        Assertions.assertThat(dmnResult.hasErrors()).isFalse();
+        final BigDecimal result = (BigDecimal) dmnResult.getDecisionResultByName(REGRESSION_DECISION_NAME).getResult();
+        Assertions.assertThat(result).isEqualTo(new BigDecimal("52.5"));
     }
 }
