+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-dmn-pmml/src/test/resources/logback-test.xml
@@ -1,0 +1,32 @@
+<configuration>
+
+  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+    <!-- encoders are assigned the type
+         ch.qos.logback.classic.encoder.PatternLayoutEncoder by default -->
+    <encoder>
+      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+    </encoder>
+  </appender>
+  
+  <appender name="KIECLIENTFILE" class="ch.qos.logback.core.FileAppender">
+    <file>target/kie-client.log</file>
+    <encoder>
+      <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+    </encoder>
+  </appender>
+
+  <logger name="org.kie.server" level="INFO" />
+  <logger name="org.apache.http" level="WARN" />
+  <logger name="org.apache.maven" level="WARN" />
+  <logger name="org.jboss.resteasy.core" level="DEBUG" />
+<!--
+  <logger name="org.kie.server.services.dmn" level="DEBUG" />
+  <logger name="org.kie.server.remote.rest.dmn" level="DEBUG" />
+  <logger name="org.kie.server.client.impl.AbstractKieServicesClientImpl" level="DEBUG">
+    <appender-ref ref="KIECLIENTFILE" />
+  </logger>
+ -->
+  <root level="INFO">
+    <appender-ref ref="STDOUT"/>
+  </root>
+</configuration>
